requirements.txt
streamlit
pandas
yfinance>=0.2.50
curl_cffi
langgraph
langchain-openai
python-dotenv